Impelemented managing articles + updates to manage users UI

+++ assets/js/application.js
@@ -19,29 +19,53 @@
       }
     });
 
-    $('#btn-generate-password').on('click', function() {
+    $('#btn-generate-password').on('click', function () {
       // TODO: Generate random value.
       $('#input-user-password').val('123456');
     });
 
-    // Manage users page
+    // Manage users [delete user]
     $('.delete-user').on('click', function (e) {
       var $this = $(this);
       var $container = $this.parents('.user-item');
       var id = $this.attr('id').replace('user-', '');
 
-      $this.button('loading');
+      if (confirm('Are you sure you want to delete this user?')) {
+        $this.button('loading');
 
-      $.ajax({
-        url: app.base_url + 'ajax/users/' + id,
-        type: 'delete',
-        dataType: 'json',
-        success: function (res) {
-          $container.fadeOut('slow').done(function () {
-            $container.remove();
-          });
-        }
-      });
+        $.ajax({
+          url: app.base_url + 'ajax/users/' + id,
+          type: 'delete',
+          dataType: 'json',
+          success: function (res) {
+            $container.fadeOut('slow').done(function () {
+              $container.remove();
+            });
+          }
+        });
+      }
+
+      return false;
+    });
+
+    // Manage articles
+    $('.delete-article').on('click', function (e) {
+      var $this = $(this);
+      var $container = $(this).parents('tr.article-item');
+      var id = $this.attr('id').replace('delete-article-', '');
+
+      if (confirm('Are you sure you want to delete this article?')) {
+        $.ajax({
+          url: app.base_url + 'ajax/articles/' + id,
+          type: 'delete',
+          dataType: 'json',
+          success: function (res) {
+            $container.fadeOut('slow').done(function() {
+              $container.remove();
+            });
+          }
+        });
+      }
 
       return false;
     });
